+++ src/components/exchange/ExchangeSearch.js
@@ -1,4 +1,5 @@
 import analytics from '@segment/analytics-react-native';
+import lang from 'i18n-js';
 import { isEmpty } from 'lodash';
 import React, { useCallback, useContext, useEffect, useRef } from 'react';
 import RadialGradient from 'react-native-radial-gradient';
@@ -120,7 +121,7 @@ const ExchangeSearch = (
     onFocus,
     searchQuery,
     testID,
-    placeholderText = 'Search Uniswap',
+    placeholderText = lang.t('button.exchange_search_uniswap'),
     clearTextOnFocus = true,
   },
   ref