+++ src/helpers/assets.js
@@ -2,7 +2,6 @@ import {
   chunk,
   compact,
   concat,
-  find,
   forEach,
   get,
   groupBy,
@@ -15,9 +14,11 @@ import {
 import { add, convertAmountToNativeDisplay, greaterThan } from './utilities';
 import store from '@rainbow-me/redux/store';
 import {
+  ETH_ADDRESS,
   ETH_ICON_URL,
   supportedNativeCurrencies,
 } from '@rainbow-me/references';
+import { ethereumUtils } from '@rainbow-me/utils';
 
 const COINS_TO_SHOW = 5;
 
@@ -36,7 +37,7 @@ const addEthPlaceholder = (
   nativeCurrency,
   emptyCollectibles
 ) => {
-  const hasEth = !!find(assets, asset => asset.address === 'eth');
+  const hasEth = !!ethereumUtils.getAccountAsset(ETH_ADDRESS);
 
   const { genericAssets } = store.getState().data;
   if (
@@ -57,7 +58,6 @@ const addEthPlaceholder = (
       icon_url: ETH_ICON_URL,
       isCoin: true,
       isPinned: pinnedCoins.includes('eth'),
-      isPlaceholder: true,
       isSmall: false,
       name: 'Ethereum',
       native: {
@@ -96,7 +96,7 @@ const getTotal = assets =>
   );
 
 export const buildCoinsList = (
-  assetsOriginal,
+  sortedAssets,
   nativeCurrency,
   isCoinListEdited,
   pinnedCoins,
@@ -110,7 +110,7 @@ export const buildCoinsList = (
     hiddenAssets = [];
 
   const { addedEth, assets } = addEthPlaceholder(
-    assetsOriginal,
+    sortedAssets,
     includePlaceholder,
     pinnedCoins,
     nativeCurrency,
@@ -193,7 +193,7 @@ export const buildCoinsList = (
 
 // TODO make it better
 export const buildBriefCoinsList = (
-  assetsOriginal,
+  sortedAssets,
   nativeCurrency,
   isCoinListEdited,
   pinnedCoins,
@@ -202,7 +202,7 @@ export const buildBriefCoinsList = (
   emptyCollectibles
 ) => {
   const { assets, smallBalancesValue, totalBalancesValue } = buildCoinsList(
-    assetsOriginal,
+    sortedAssets,
     nativeCurrency,
     isCoinListEdited,
     pinnedCoins,
@@ -328,16 +328,10 @@ export const buildBriefUniqueTokenList = (
   uniqueTokens,
   selectedShowcaseTokens
 ) => {
-  const uniqueTokensNotInShowcase = uniqueTokens.filter(
-    ({ uniqueId }) => !selectedShowcaseTokens.includes(uniqueId)
-  );
   const uniqueTokensInShowcase = uniqueTokens
     .filter(({ uniqueId }) => selectedShowcaseTokens.includes(uniqueId))
     .map(({ uniqueId }) => uniqueId);
-  const grouped2 = groupBy(
-    uniqueTokensNotInShowcase,
-    token => token.familyName
-  );
+  const grouped2 = groupBy(uniqueTokens, token => token.familyName);
   const families2 = sortBy(Object.keys(grouped2), row =>
     row.replace(regex, '').toLowerCase()
   );