+++ src/hooks/useAccountEmptyState.js
@@ -11,6 +11,7 @@ export default function useAccountEmptyState() {
   useEffect(() => {
     const checkStorage = async () => {
       const reallyEmpty = await getAccountEmptyState(accountAddress, network);
+
       if (reallyEmpty) {
         setIsReallyEmpty(true);
       } else {