+++ src/screens/TransactionConfirmationScreen.js
@@ -57,7 +57,6 @@ import { isDappAuthenticated } from '@rainbow-me/helpers/dappNameHandler';
 import { findWalletWithAccount } from '@rainbow-me/helpers/findWalletWithAccount';
 import networkTypes from '@rainbow-me/helpers/networkTypes';
 import {
-  useAccountAssets,
   useAccountSettings,
   useCurrentNonce,
   useDimensions,
@@ -94,11 +93,13 @@ import {
   isMessageDisplayType,
   isSignFirstParamType,
   isSignSecondParamType,
+  isSignTypedData,
   isTransactionDisplayType,
   PERSONAL_SIGN,
   SEND_TRANSACTION,
   SIGN,
   SIGN_TYPED_DATA,
+  SIGN_TYPED_DATA_V4,
 } from '@rainbow-me/utils/signingMethods';
 import logger from 'logger';
 
@@ -155,7 +156,6 @@ const NOOP = () => undefined;
 
 export default function TransactionConfirmationScreen() {
   const { colors } = useTheme();
-  const { allAssets } = useAccountAssets();
   const [provider, setProvider] = useState();
   const [currentNetwork, setCurrentNetwork] = useState();
   const [isAuthorizing, setIsAuthorizing] = useState(false);
@@ -168,7 +168,6 @@ export default function TransactionConfirmationScreen() {
   const { wallets, walletNames, switchToWalletWithAddress } = useWallets();
   const balances = useWalletBalances(wallets);
   const { accountAddress, nativeCurrency } = useAccountSettings();
-  const getNextNonce = useCurrentNonce(accountAddress, currentNetwork);
   const keyboardHeight = useKeyboardHeight();
   const dispatch = useDispatch();
   const { params: routeParams } = useRoute();
@@ -221,6 +220,8 @@ export default function TransactionConfirmationScreen() {
     };
   }, [currentNetwork, walletConnector?._accounts, walletNames, wallets]);
 
+  const getNextNonce = useCurrentNonce(accountInfo.address, currentNetwork);
+
   const isL2 = useMemo(() => {
     return isL2Network(currentNetwork);
   }, [currentNetwork]);
@@ -250,11 +251,12 @@ export default function TransactionConfirmationScreen() {
       setNativeAsset(asset);
     };
     currentNetwork && getNativeAsset();
-  }, [accountInfo.address, allAssets, currentNetwork]);
+  }, [accountInfo.address, currentNetwork]);
 
   const {
     gasLimit,
     isSufficientGas,
+    isValidGas,
     startPollingGasFees,
     stopPollingGasFees,
     updateGasFeeOption,
@@ -273,7 +275,7 @@ export default function TransactionConfirmationScreen() {
       isSufficientGasChecked
     )
       return;
-    updateGasFeeOption(selectedGasFeeOption, [nativeAsset]);
+    updateGasFeeOption(selectedGasFeeOption);
     setIsSufficientGasChecked(true);
   }, [
     isSufficientGas,
@@ -541,7 +543,6 @@ export default function TransactionConfirmationScreen() {
 
     setIsBalanceEnough(isEnough);
   }, [
-    allAssets,
     isBalanceEnough,
     isMessageRequest,
     isSufficientGas,
@@ -751,6 +752,7 @@ export default function TransactionConfirmationScreen() {
       case PERSONAL_SIGN:
         response = await signPersonalMessage(message, existingWallet);
         break;
+      case SIGN_TYPED_DATA_V4:
       case SIGN_TYPED_DATA:
         response = await signTypedDataMessage(message, existingWallet);
         break;
@@ -790,13 +792,14 @@ export default function TransactionConfirmationScreen() {
     if (isMessageRequest) {
       return handleSignMessage();
     }
-    if (!isBalanceEnough) return;
+    if (!isBalanceEnough || !isValidGas) return;
     return handleConfirmTransaction();
   }, [
     handleConfirmTransaction,
     handleSignMessage,
     isBalanceEnough,
     isMessageRequest,
+    isValidGas,
   ]);
 
   const onPressSend = useCallback(async () => {
@@ -819,13 +822,17 @@ export default function TransactionConfirmationScreen() {
       ready = false;
     }
     return !isMessage &&
-      isBalanceEnough === false &&
+      (isBalanceEnough === false || !isValidGas) &&
       isSufficientGas !== null ? (
       <Column marginBottom={24} marginTop={19}>
         <SheetActionButton
           color={colors.transparent}
           disabled
-          label={`${nativeAsset?.symbol} balance too low`}
+          label={
+            !isValidGas
+              ? `Invalid fee`
+              : `${nativeAsset?.symbol} balance too low`
+          }
           onPress={onCancel}
           size="big"
           textColor={colors.avatarColor[7]}
@@ -853,14 +860,15 @@ export default function TransactionConfirmationScreen() {
       </SheetActionButtonRow>
     );
   }, [
-    colors,
-    isBalanceEnough,
     isMessageRequest,
+    isBalanceEnough,
     isSufficientGas,
+    isValidGas,
+    colors,
     nativeAsset?.symbol,
     onCancel,
-    onPressSend,
     onPressCancel,
+    onPressSend,
   ]);
 
   const renderTransactionSection = useCallback(() => {
@@ -930,7 +938,7 @@ export default function TransactionConfirmationScreen() {
   const ShortSheetHeight = 486 + safeAreaInsetValues.bottom;
   const TallSheetHeight = 656 + safeAreaInsetValues.bottom;
   const MessageSheetHeight =
-    (method === SIGN_TYPED_DATA ? 630 : android ? 595 : 580) +
+    (isSignTypedData(method) ? 630 : android ? 595 : 580) +
     safeAreaInsetValues.bottom;
 
   const balanceTooLow = isBalanceEnough === false && isSufficientGas !== null;